Add build target classes to Fab
This adds a module directory containing build target classes.  The
base module contains an abstract build implementation which is
intended to be subclassed by the application to create a set of rules
which can be used to compile an application.

It also creates a zero-config class derived from the base class which
is intended to show how the rules should be implemented.