Revised fab command line argument parser
Based on discussions, the fab command line argument parser needs to
support two phases.  In the first phase, it should determine the name
of the file containing the build target rules.  This can be explicitly
specified with a --file option, or it can check a default location, or
the file can be missing entirely in which case an internal zero-config
target should be used.

The second phase of parsing uses an argparse.ArgumentParser subclass.
This can be configured with add_argument() in the usual way.

When either parse_args() or parse_known_args() is called, a decorator
checks for some standard arguments and adds them if they do not exist.
This allows the caller to specify alternate versions of these routines
if they need to.  After the arguments have been parsed, the decorator
performs some additional checking and setup work, e.g. configuring the
logging framework.