Revised fab command line argument parser
Two-phase command line argument parser based on argparse.  The first
phase locates the Fab target file and sets zero config mode if the
file is not found.  This allows the target to set additional command
line options.  The second phases parses the full set of arguments as
normal.

The parser adds a number of default options if these have not been
added by the caller.