Fix a design flaw and address review issues
This adds a quiet mode test based on reviewer's suggstion to use a
StringIO instance as the output stream.  It also adds some missing
items to the file logging test.

In the process of implementing the file test, a flaw was spotted in
the existing implemenation of setup_logging() which resulted in log
messages being excluded from both the console stream (intended) and
the log file (not intended).

This problem was fixed by creating a FabLogFilter class and using
this, rather than the logLevel of each hierarchical logger to control
whether the message gets sent to the console stream.  This allows the
different log handlers to be managed separately.

Other fixes include the addition of missing type hinting, improvements
to docstrings, and the addition of a check in make_logger() to ensure
the module is defined - both for correctness and to prevent an error
from type type checking.