Fix a bug in and improve the test unit tests
In response to comments in review, this improves one of the checks in
the logtools module to allow a value of None or zero.

This also enhances the unit tests for logtools by extend the range of
log messages and improves the checks on the output rather than simply
counting the number of lines.  It also adds a test to check that the
message formats change to provide more details when system logging is
enabled.