Working directory should default to current directory (#457)
At the moment the fab workspace is created in the root of the users home
directory. A more suitable place for the fab workspace to default to
would be the current working directory. This pull request aims to
implement a fix to this as outlined by
https://github.com/MetOffice/fab/issues/338.

This PR is the replacement PR for #453 in the hope that commit histories
will be slightly tidier in future PRs. Once merged this pull request
closes #338 .