Add ability to cache command line arguments
This adds a feature which allows complex command line arguments to be
written to a cache file and then re-used.  Where an argument is
specified in both the cache and on the current command line, the
command line version takes priority.

Cache arguments themselves are not sticky, so if --save-cache and
--use-cache are set on the command line, they will get written to the
cache but ignored when the cache is next loaded.  This prevents the
cache from being inadvertently overwritten with new arguments.