Move more caching features into the mixin
Some of the caching features were defined in the wrong places; the
argument group was added in FabArgumentParser and the post-parse_args
setup was entirely contained in the _parser_wrapper decorator.  This
prevents the class from being self-contained and makes it harder to
reuse.  This moves the methods into the _FabArgumentCache class.

This also creates a CachingArgumentParser which is a simple
combination of _FabArgumentCache and ArgumentParser.  This provides a
self-contained parser for use by infrastructure that does not use the
fab CUI.