Fix output_fpath being cast to string in archive_objects (#472)
During https://github.com/MetOffice/fab/pull/431 implementation, it was
found that output_fpath is always passed to Ar.create as a string,
despite being type hinted as a Path. Currently, this requires creating a
new Path instance from output_fpath to unlink archive (.a) files when
Ar.create is called . Replacing the casting with a direct Path
instantiation in archive_objects.py (lines 128-129) would be ideal, but
since it involves usage of a Template instance, we need to verify that
this change doesn't affect program semantics.

Once this PR is merged it shall close
https://github.com/MetOffice/fab/issues/441 and will also close
https://github.com/MetOffice/fab/issues/359.

This PR replaces #459 which I accidentally closed.