Fix problems with CI testing
There appear to be two problems with the github continuous integration
tests.

The format of the argparse help for choices is quoted in some
versions of python, so omit the choice strings from the output check.

The git system tests need details of the current user to run.  Add a
step which adds some temporary user details.